+++ code/go/site/routes_examples_dispatch_custom_event.go
@@ -40,7 +40,7 @@ document.addEventListener('example-event-from-server', (e) => {
 					return
 				}
 
-				sse.DispatchCustomEventMarshal("example-event-from-server", detail)
+				sse.DispatchCustomEvent("example-event-from-server", detail)
 			}
 		}
 	})